Added support for applying borders in parallel
Uses multiprocessing under the hood. Default is to not do anything in parallel. This update required changing some of the internal APIs (_add_border now takes a file path as input and returns a file path as output rather than passing aroung PIL.Image objects). This was necessary to work with multiprocessing (pickling those images directly would be stupid), but the result is quite neat.